+++ NEWS.md
@@ -1,7 +1,9 @@
 # riekelib (development version)
 
 * Change default backend for plog rendering to [ragg](https://ragg.r-lib.org/). 
-* Add `cholesky_decompose()`
+* Add functions for working with Gaussian processes:
+  * `cholesky_decompose()`: generate the lower triangular Cholesky decomposition of a matrix.
+  * `cov_exp_quad()`: generate a covariance matrix using the exponentiated quadratic kernel. 
 
 # riekelib 0.1.0